Explicitly initialize plugins (fixes #38)
This allows a user to import plugins on local development environments
which do not have access to AWS metadata without stalling the process
at boot time.

Each of the three plugins will need to be imported and initialized manually.
The API will safeguard against multiple requests but it's not goroutine safe.