Synchronize log messages in default logger.
Use a mutex to synchronize calls to underling io.Writer. This avoids
log message interleaving and potentially worse race conditions,
depending on what the writer is.

Also tweak the invalid LogLevel String() to be UNKNOWNLOGLEVEL instead
of just UNKNOWN so it is easier to understand what is unknown if it
ever shows up.