Logging interface (#82)
* Replace seelog logger with internal logger.

The seelog logger had various drawbacks:
- not easily user configurable (e.g. can't disable, can't write to a
  file, can't write to stderr, can't write to json, etc)
- uses global seelog instance which can conflict w/ other packages
  or user's logger

The replacement logger uses a simple Logger interface:

type Logger interface {
	Log(level LogLevel, msg fmt.Stringer)
}

Rather than a separate method for each log level, the level is passed
as an argument. This makes it easy/possible to add more log levels in
the future. The lowest log level is now "debug" instead of seelog's
"trace" (seemed mildly confusing to have a log method named "trace" in
a cloud tracing package).

We use a fmt.Stringer so any log message serialization can be deferred
until the message is actually written.

There is a default Logger implementation the user can use via
NewDefaultLogger(io.Writer, LogLevel) to create a logger that writes
log messages of at least a certain log level to a specified io.Writer.

The logger defaults to a stdout writer of min level "info".

The "xraylog" package is the user visible logging API. It contains
"xray" in the package name so it doesn't annoyingly look like every
other "logger" package (e.g. to goimports). The user can set a new
logger using xray.SetLogger().

The actual logging instance and logging wrapper methods are in an
internal "logger" package hidden from the user. There are
Info(...interface{}) and Infof(string, ...interface{}) varieties for
each log level. In addition there is a DebugDeferred(func() string)
takes an arbitrary function (which won't get called unless the log
message is actually logged).

* Add README.md note about logger interface

* Synchronize log messages in default logger.

Use a mutex to synchronize calls to underling io.Writer. This avoids
log message interleaving and potentially worse race conditions,
depending on what the writer is.

Also tweak the invalid LogLevel String() to be UNKNOWNLOGLEVEL instead
of just UNKNOWN so it is easier to understand what is unknown if it
ever shows up.