Bugfix (#115)
* Fix race bug with err

The err is otherwise in the scope of the returned error
and a race with the spawned goroutine

* Fix race bugs with ClientTrace

None of the callbacks of ClientTrace can be assumed to not
call in goroutines with other callbacks.  We at least
need to lock each of these functions.  It is possible
we could investigate further and reduce some of these locks
once we clean up the segment locking order